requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
uvicorn==0.38.0
gunicorn==23.0.0
python-dotenv==1.0.1
dj-database-url==2.1.0

# Wallet dependencies
bip-utils>=2.9.0  # For BIP84 native segwit derivation (primary)
hdwallet>=2.2.1  # HD wallet address derivation (fallback)
base58>=2.1.1  # For zpub to xpub conversion (required for BIP84)
django-redis>=5.4.0  # For Redis cache backend
cryptography>=41.0.0  # For private key encryption (AES-256)
bit>=0.8.0  # For Bitcoin transaction creation and broadcasting
pycryptodome>=3.19.0  # For Keccak-256 hash (Ethereum address derivation)
